+++ src/var_table.c
@@ -22,9 +22,10 @@ Nst_VarTable *Nst_vt_new(Nst_MapObj *global_table, Nst_SeqObj *args,
 #ifdef _Nst_ARCH_x64
     Nst_Obj *debug_str = Nst_str_new_c_raw("x64", false);
 #else
-    Nst_Obj *debug_str = Nst_str_new_c_raw("", false);
+    Nst_Obj *debug_str = Nst_str_new_c_raw("x86", false);
 #endif // !_Nst_ARCH_x64
-    Nst_map_set_str(vars, "_debug_", debug_str);
+    Nst_map_set_str(vars, "_debug_", Nst_c.Bool_true);
+    Nst_map_set_str(vars, "_debug_arch_", debug_str);
     Nst_dec_ref(debug_str);
 #endif // !_DEBUG